requirements.txt
discord
asyncio
colorama
python-dotenv
Werkzeug
Flask
prettytable